services/user_service/requirements.txt
grpcio
grpcio-tools
grpcio-reflection
firebase-admin
requests
protobuf

